Refactor: 보안 관련 클래스 디렉토리 이동 및 중앙화된 사용자 인증을 위한 SecurityUtil 추가
- 현재 사용자 정보 조회를 위한 SecurityUtil 클래스 구현
- 관련 서비스에 SecurityUtil 통합(DailyLook)
